+++ frontend/src/components/NewChecklist/NewRoomInfoForm/useDefaultRoomName.ts
@@ -1,0 +1,23 @@
+import { useEffect } from 'react';
+import { useStore } from 'zustand';
+
+import useGetChecklistListQuery from '@/hooks/query/useGetChecklistListQuery';
+import checklistRoomInfoStore, { initialRoomInfo } from '@/store/checklistRoomInfoStore';
+
+const useDefaultRoomName = () => {
+  const actions = useStore(checklistRoomInfoStore, state => state.actions);
+  const roomName = useStore(checklistRoomInfoStore, state => state.rawValue.roomName);
+
+  const { data: checklistList } = useGetChecklistListQuery();
+  useEffect(() => {
+    if (!checklistList) return;
+    const count = checklistList.filter(
+      checklist => new Date(checklist.createdAt).getUTCDay() === new Date().getUTCDay(),
+    ).length;
+    if (roomName !== initialRoomInfo.roomName) return;
+    const date = new Date();
+    actions.set('roomName', `${date.getMonth()}월 ${date.getDay()}일 ${count + 1}번째 방`);
+  }, [checklistList]);
+};
+
+export default useDefaultRoomName;
